test: moved service mocks into own test file
Moved any mock services used in testing into their own mock_test.go file to keep them separated from other tests (was getting too large)
